Move funFact to FunFact component
